--- src/styles/components/atoms/Member.module.scss
@@ -1,47 +1,0 @@
-.member_box {
-  .member {
-    display: flex;
-    width: 240px;
-    padding: 5px;
-
-    .avatar {
-      width: 48px;
-      height: 48px;
-      border-radius: 24px;
-    }
-
-    .name {
-      margin: auto 3px auto 10px;
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      font-size: 20px;
-    }
-
-    .bot_name {
-      margin: auto 0 auto 10px;
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      font-size: 20px;
-    }
-
-    .bot {
-      margin: auto 0 auto 5px;
-      font-size: 14px;
-      color: #ffffff;
-      background-color: navy;
-      padding: 3px 5px;
-      border-radius: 3px;
-    }
-  }
-
-  .membercard {
-    transform: scale(0);
-    transition: transform 0.2s;
-
-    &.show {
-      transform: translateX(-105%);
-    }
-  }
-}
